java/spellcheck-javadocs/requirements.txt
javalang
pyenchant
bs4
